requirements.txt
requests
boto3
Pillow